refactor: Repository 스펠링 고침
